requirements.txt
numpy~=1.26.4
pandas~=2.2.2
simpletransformers~=0.70.0
opencv-python~=4.9.0.80
fastapi~=0.110.2
python-multipart~=0.0.9
uvicorn~=0.29.0
transformers~=4.39.3
pytesseract~=0.3.10
scikit-learn~=1.4.2
APScheduler~=3.10.4
matplotlib~=3.8.4
pytest~=8.2.1
accelerate~=0.31.0
jiwer~=3.0.4
datasets~=2.18.0
python-dateutil~=2.9.0.post0